backend/requirements.txt
# Flask backend requirements
flask
flask-cors
pymongo
python-dotenv
